+++ .github/workflows/go.yml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        go: ["1.18.x", "1.19.x"]
+        go: ["1.17.x", "1.18.x", "1.19.x"]
         include:
         - go: 1.19.x
           os: "ubuntu-latest"
